backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-admin-sortable2==2.2.4
django-jazzmin==3.0.1
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
